placing the pieces works pending Unit test.
I suspet a point of failure in the ArrayLists since I think it's possible to have the same index in two ArrayList. I'm think of transfering everything into a HashSet before flipping the pieces. But I don't want to do something I don't have to so I'm waiting AFTER the Unit test to correct his.